refactored components in transaction page
